Move styled components to styles
